pyproject.toml
[project]
name = "heb-sentence-gen-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.101.0",
    "phonikud-onnx>=1.0.4",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
